+++ tests/bfabric_app_runner/app_runner/test_resolve_app.py
@@ -1,5 +1,4 @@
-import importlib.resources
-from pathlib import Path
+from pathlib import Path, PosixPath
 
 import pytest
 import yaml
@@ -10,56 +9,46 @@ from bfabric.experimental.workunit_definition import (
     WorkunitRegistrationDefinition,
 )
 from bfabric_app_runner.app_runner.resolve_app import load_workunit_information
+from bfabric_app_runner.specs.app.app_spec import AppSpec, BfabricAppSpec
 from bfabric_app_runner.specs.app.app_version import AppVersion
 from bfabric_app_runner.specs.app.commands_spec import CommandsSpec, CommandExec
 
 
 class TestLoadWorkunitInformation:
     @pytest.fixture
-    def app_version(self):
-        return AppVersion(
-            version="xyz",
-            commands=CommandsSpec(
-                dispatch=CommandExec(
-                    command="bash -c 'echo dispatching'",
-                ),
-                process=CommandExec(
-                    command="bash -c 'echo processing'",
-                ),
-            ),
+    def app_spec(self):
+        return AppSpec(
+            bfabric=BfabricAppSpec(app_runner="0.1.0"),
+            versions=[
+                AppVersion(
+                    version="xyz",
+                    commands=CommandsSpec(
+                        dispatch=CommandExec(
+                            command="bash -c 'echo dispatching'",
+                        ),
+                        process=CommandExec(
+                            command="bash -c 'echo processing'",
+                        ),
+                    ),
+                )
+            ],
         )
 
     @pytest.fixture
-    def app_version_yaml(self, app_version) -> str:
-        return yaml.safe_dump(app_version.model_dump(mode="json"))
+    def app_spec_yaml(self, app_spec) -> str:
+        return yaml.safe_dump(app_spec.model_dump(mode="json"))
 
     @pytest.fixture
-    def app_version_path(self, fs, app_version_yaml):
-        path = Path("/fake/work/app_version.yml")
-        fs.create_file(path, contents=app_version_yaml)
+    def app_spec_path(self, fs, app_spec_yaml):
+        path = Path("/fake/work/app_spec.yml")
+        fs.create_file(path, contents=app_spec_yaml)
         return path
-
-    @pytest.fixture
-    def app_spec_module(self, mocker, app_version_yaml) -> str:
-        unmocked_read_text = importlib.resources.read_text
-
-        def mocked_read_text(*args, **kwargs):
-            if args[0] == "_mocked_package.integrations.bfabric":
-                if args[1] == "app.yml":
-                    return app_version_yaml
-                else:
-                    raise NotImplementedError
-            else:
-                return unmocked_read_text(*args, **kwargs)
-
-        mocker.patch("importlib.resources.read_text").side_effect = mocked_read_text
-        return "_mocked_package"
 
     @pytest.fixture
     def workunit_definition(self):
         return WorkunitDefinition(
             execution=WorkunitExecutionDefinition(
-                raw_parameters={"hello": "world"},
+                raw_parameters={"hello": "world", "application_version": "xyz"},
                 dataset=1234,
             ),
             registration=WorkunitRegistrationDefinition(
@@ -70,7 +59,7 @@ class TestLoadWorkunitInformation:
                 container_id=3,
                 container_type="project",
                 storage_id=4,
-                storage_output_folder=Path("/fake/remote/path"),
+                storage_output_folder=PosixPath("/fake/remote/path"),
             ),
         )
 
@@ -84,35 +73,20 @@ class TestLoadWorkunitInformation:
         fs.create_file(path, contents=workunit_definition_yaml)
         return path
 
-    def test_from_module_when_default(self, mocker, app_spec_module, workunit_definition_path, app_version):
+    def test_from_path(self, mocker, app_spec_path, workunit_definition_path, app_spec, workunit_definition):
         mock_client = mocker.Mock(name="mock_client")
-        return_app_version, return_workunit_ref = load_workunit_information(
-            app_spec=app_spec_module,
-            client=mock_client,
-            work_dir=Path("/fake/work"),
-            workunit_ref=workunit_definition_path,
+        mocker.patch(
+            "bfabric_app_runner.app_runner.resolve_app._resolve_workunit_definition",
+            return_value=(workunit_definition, workunit_definition_path),
         )
-        assert return_app_version == app_version
-        assert return_workunit_ref == workunit_definition_path
 
-    def test_from_module_when_explicit(self, mocker, app_spec_module, workunit_definition_path, app_version):
-        mock_client = mocker.Mock(name="mock_client")
-        return_app_version, return_workunit_ref = load_workunit_information(
-            app_spec=f"{app_spec_module}.integrations.bfabric:app.yml",
+        return_app_version, return_bfabric_app_spec, return_workunit_ref = load_workunit_information(
+            app_spec=app_spec_path,
             client=mock_client,
             work_dir=Path("/fake/work"),
             workunit_ref=workunit_definition_path,
         )
-        assert return_app_version == app_version
-        assert return_workunit_ref == workunit_definition_path
 
-    def test_from_path(self, mocker, app_version_path, workunit_definition_path, app_version):
-        mock_client = mocker.Mock(name="mock_client")
-        return_app_version, return_workunit_ref = load_workunit_information(
-            app_spec=app_version_path,
-            client=mock_client,
-            work_dir=Path("/fake/work"),
-            workunit_ref=workunit_definition_path,
-        )
-        assert return_app_version == app_version
+        assert return_app_version == app_spec.versions[0]
+        assert return_bfabric_app_spec == app_spec.bfabric
         assert return_workunit_ref == workunit_definition_path