+++ tests/bfabric_app_runner/integration/test_cli_commands.py
@@ -8,13 +8,6 @@ import pytest
     "command",
     [
         "--help",
-        "app --help",
-        "app dispatch --help",
-        "app run --help",
-        "chunk --help",
-        "chunk outputs --help",
-        "chunk process --help",
-        "chunk run-all --help",
         "inputs --help",
         "inputs check --help",
         "inputs clean --help",